+++ morpheus-graphql-core/src/Data/Morpheus/Parsing/Internal/Pattern.hs
@@ -10,6 +10,8 @@ module Data.Morpheus.Parsing.Internal.Pattern
     enumValueDefinition,
     inputFieldsDefinition,
     parseOperationType,
+    argumentsDefinition,
+    parseDirectiveLocation,
   )
 where
 
@@ -41,6 +43,7 @@ import Data.Morpheus.Types.Internal.AST
   ( ArgumentsDefinition (..),
     DataEnumValue (..),
     Directive (..),
+    DirectiveLocation (..),
     FieldContent (..),
     FieldDefinition (..),
     FieldName,
@@ -52,8 +55,10 @@ import Data.Morpheus.Types.Internal.AST
     TypeName,
     Value,
   )
+import Data.Text (pack)
 import Text.Megaparsec
   ( (<|>),
+    choice,
     label,
     many,
     optional,
@@ -174,3 +179,34 @@ parseOperationType = label "OperationType" $ do
       <|> (string "subscription" $> Subscription)
   ignoredTokens
   return kind
+
+parseDirectiveLocation :: Parser DirectiveLocation
+parseDirectiveLocation =
+  label
+    "DirectiveLocation"
+    ( choice $
+        toKeyword
+          <$> [ FIELD_DEFINITION,
+                FRAGMENT_DEFINITION,
+                FRAGMENT_SPREAD,
+                INLINE_FRAGMENT,
+                ARGUMENT_DEFINITION,
+                INTERFACE,
+                ENUM_VALUE,
+                INPUT_OBJECT,
+                INPUT_FIELD_DEFINITION,
+                SCHEMA,
+                SCALAR,
+                OBJECT,
+                QUERY,
+                MUTATION,
+                SUBSCRIPTION,
+                UNION,
+                ENUM,
+                FIELD
+              ]
+    )
+    <* ignoredTokens
+
+toKeyword :: Show a => a -> Parser a
+toKeyword x = string (pack $ show x) $> x