+++ morpheus-graphql-core/src/Data/Morpheus/Schema/Schema.hs
@@ -3,77 +3,60 @@
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE QuasiQuotes #-}
+{-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 
 module Data.Morpheus.Schema.Schema
-  ( withSystemTypes,
-    systemTypes,
+  ( internalSchema,
   )
 where
 
 -- MORPHEUS
 
-import Data.Morpheus.Internal.Utils
-  ( (<:>),
-    Failure (..),
-    resolveUpdates,
-    singleton,
-  )
 import Data.Morpheus.Schema.DSL (dsl)
 import Data.Morpheus.Types.Internal.AST
-  ( ANY,
-    DataFingerprint (..),
-    FieldsDefinition,
-    GQLErrors,
-    Message,
-    OUT,
+  ( DataFingerprint (..),
     Schema (..),
-    TypeContent (..),
     TypeDefinition (..),
-    TypeWrapper (..),
-    insertType,
     internalFingerprint,
-    mkInputValue,
-    mkObjectField,
-    unsafeFromFields,
   )
 
-withSystemTypes :: (Monad m, Failure GQLErrors m, Failure Message m) => Schema s -> m (Schema s)
-withSystemTypes s@Schema {query = q@TypeDefinition {typeContent = DataObject inter fields}} =
-  ( do
-      fs <- fields <:> hiddenFields
-      pure $ s {query = q {typeContent = DataObject inter fs}}
-  )
-    >>= (`resolveUpdates` map (insertType . internalType) systemTypes)
-withSystemTypes _ = failure ("Query must be an Object Type" :: Message)
+toInternalSchema :: Schema s -> Schema s
+toInternalSchema Schema {..} = Schema {types = fmap toInternalType types, ..}
 
-hiddenFields :: FieldsDefinition OUT s
-hiddenFields =
-  unsafeFromFields
-    [ mkObjectField
-        (singleton (mkInputValue "name" [] "String"))
-        "__type"
-        [TypeMaybe]
-        "__Type",
-      mkInputValue
-        "__schema"
-        []
-        "__Schema"
-    ]
-
-internalType :: TypeDefinition a s -> TypeDefinition a s
-internalType
+toInternalType :: TypeDefinition a s -> TypeDefinition a s
+toInternalType
   tyDef@TypeDefinition
     { typeFingerprint = DataFingerprint name xs
     } =
     tyDef {typeFingerprint = internalFingerprint name xs}
 
-systemTypes :: [TypeDefinition ANY s]
-systemTypes =
-  [dsl|
+internalSchema :: Schema s
+internalSchema =
+  toInternalSchema
+    [dsl|
+
+"""
+Directs the executor to skip this field or fragment when the `if` argument is true.
+"""
+directive @skip(if: Boolean!) 
+  on FIELD | FRAGMENT_SPREAD | INLINE_FRAGMENT
+
+"""
+Directs the executor to include this field or fragment only when the `if` argument is true.
+"""
+directive @include(if: Boolean!) 
+  on FIELD | FRAGMENT_SPREAD | INLINE_FRAGMENT
+
+"""
+Marks an element of a GraphQL schema as no longer supported.
+"""
+directive @deprecated(reason: String) 
+  on FIELD_DEFINITION | ENUM_VALUE
+
 
 scalar Boolean
 scalar Int
@@ -173,6 +156,11 @@ enum __TypeKind {
   INPUT_OBJECT
   LIST
   NON_NULL
+}
+
+type Query {
+  __type(name: String!): __Type
+  __schema: __Schema!
 }
 
 |]